Add Permission Requirement for .Spec
This allows server owners to restrict ghosting to specific roles. This way, server owners could allow patrons, or moderators only to spectate. If server owners want to allow everyone to use the command, they can give this permission to the default permission group, allowing everyone to use this command. This could also be implemented with a config option boolean that allows everyone to use the command when true, but restricts it to only users with this permission when false, preventing server owners from needing to add this permission at all if it won't be used.